+++ src/statusBar/AbstractHandler.ts
@@ -10,9 +10,28 @@ export abstract class AbstractHandler {
             || (!this.isEnabledFor(editor.document.languageId))) {
             return undefined;
         }
-        return editor.selection.isEmpty
-            ? editor.document.lineAt(editor.selection.active.line).text
-            : editor.document.getText(editor.selection);
+
+        // No selection; uses cursor position
+        if (editor.selection.isEmpty) {
+            return editor.document.lineAt(editor.selection.active.line).text;
+        }
+
+        // Do not expand selection
+        if (!config.expandSelectionToLine) {
+            return editor.document.getText(editor.selection);
+        }
+
+        // Expand single line selection to line
+        if (editor.selection.isSingleLine) {
+            return editor.document.lineAt(editor.selection.start.line).text;
+        }
+
+        // Expand multiline selection
+        return editor.document.getText(new vscode.Range(
+            editor.selection.start.line, 0,
+            editor.selection.end.character
+                ? editor.selection.end.line + 1
+                : editor.selection.end.line, 0));
     }
 
     private isEnabledFor(languageId: string): boolean {