+++ src/statusBar/StatusBarHandler.ts
@@ -1,13 +1,13 @@
 import * as vscode from 'vscode';
 import { config } from "../config";
 import { mainParser } from "../parser/MainParser";
-import { atExitTotalTiming } from '../totalTiming/AtExitTotalTiming';
+import { AtExitTotalTiminsMeterable, atExitTotalTiming } from '../totalTiming/AtExitTotalTiming';
 import { defaultTotalTiming } from '../totalTiming/DefaultTotalTiming';
 import { executionFlowTotalTiming } from '../totalTiming/ExecutionFlowTotalTiming';
 import { TotalTimingMeterable } from '../totalTiming/TotalTiming';
 import { humanReadableBytes } from '../utils/ByteUtils';
 import { humanReadableInstructions } from "../utils/InstructionUtils";
-import { hashCode } from "../utils/TextUtils";
+import { espaceIfNotInfix, hashCode, pluralize } from "../utils/TextUtils";
 import { humanReadableTiming, humanReadableTimings } from '../utils/TimingUtils';
 import { AbstractHandler } from "./AbstractHandler";
 import { CommandHandler } from "./CommandHandler";
@@ -163,32 +163,30 @@ export class StatusBarHandler extends AbstractHandler {
     private buildText(
         defaultTiming: TotalTimingMeterable,
         flowTiming: TotalTimingMeterable | undefined,
-        atExitTiming: TotalTimingMeterable | undefined): string {
+        atExitTiming: AtExitTotalTiminsMeterable | undefined): string {
 
         // Builds the statur bar text
         let text = "";
 
         if (config.statusBar.showInstruction) {
-            const instructionIcon = config.statusBar.instructionIcon;
             const instruction = humanReadableInstructions(defaultTiming);
             if (instruction) {
-                text += `${instructionIcon} ${instruction} `;
+                const instructionIcon = espaceIfNotInfix(config.statusBar.instructionIcon);
+                text += `${instructionIcon}${instruction}`;
             }
         }
 
         const timing = this.buidTimingsText(defaultTiming, flowTiming, atExitTiming);
         if (timing) {
-            const timingsIcon = config.statusBar.timingsIcon;
-            text += `${timingsIcon} ${timing}`;
+            const timingsIcon = espaceIfNotInfix(config.statusBar.timingsIcon);
+            text += `${timingsIcon}${timing}`;
         }
 
         const size = defaultTiming.size;
         if (size) {
-            const sizeIcon = config.statusBar.sizeIcon;
-            const sizeSuffix = (config.statusBar.compactSize) ? "B"
-                : (size === 1) ? " byte"
-                    : " bytes";
-            text += ` ${sizeIcon} ${size}${sizeSuffix}`;
+            const sizeIcon = espaceIfNotInfix(config.statusBar.sizeIcon);
+            const sizeSuffix = pluralize(config.statusBar.sizeSuffix, size);
+            text += `${sizeIcon}${size}${sizeSuffix}`;
             if (config.statusBar.showBytes) {
                 const bytes = humanReadableBytes(defaultTiming);
                 if (bytes) {
@@ -197,30 +195,39 @@ export class StatusBarHandler extends AbstractHandler {
             }
         }
 
-        return text;
+        return text.trim().replace(/\s+/, " ");
     }
 
     private buidTimingsText(
         defaultTiming: TotalTimingMeterable,
         flowTiming: TotalTimingMeterable | undefined,
-        atExitTiming: TotalTimingMeterable | undefined): string | undefined {
+        atExitTiming: AtExitTotalTiminsMeterable | undefined): string | undefined {
+
+        // Applies requested order
+        const totalTimingsOrder = config.statusBar.totalTimingsOrder;
+        const [retTiming, jumpCallTiming] = atExitTiming?.isLastInstructionRet
+            ? [atExitTiming, undefined]
+            : [undefined, atExitTiming];
+        const [b, c, d] =
+            totalTimingsOrder === "retFlowJumpCall" ? [retTiming, flowTiming, jumpCallTiming]
+            : totalTimingsOrder === "flowRetJumpCall" ? [flowTiming, retTiming, jumpCallTiming]
+            : totalTimingsOrder === "retJumpCallFlow" ? [retTiming, jumpCallTiming, flowTiming]
+            : [undefined, undefined, undefined]; // (should never happen)
 
         switch (config.statusBar.totalTimings) {
             case "all":
             case "combineAll":
                 return humanReadableTimings(
-                    [defaultTiming, flowTiming, atExitTiming],
-                    config.statusBar.totalTimingsCombined);
+                    [defaultTiming, b, c, d], config.statusBar.totalTimingsCombined);
 
             case "smart":
             case "combineSmart":
-                return humanReadableTimings(
-                    [atExitTiming || flowTiming ? undefined : defaultTiming, flowTiming, atExitTiming],
-                    config.statusBar.totalTimingsCombined);
+                return flowTiming || atExitTiming
+                    ? humanReadableTimings([b, c, d], config.statusBar.totalTimingsCombined)
+                    : humanReadableTimings([defaultTiming]);
 
             case "best":
-                return humanReadableTimings(
-                    [atExitTiming || flowTiming || defaultTiming]);
+                return humanReadableTimings([atExitTiming || flowTiming || defaultTiming]);
 
             case "default":
             default: