+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 5.2.0
+
+- Ability to expand selection to line
+- More flexible status bar customization
+- Fixes settings default values being ignored. #145
+- Further division of total timing at exit based on the last instruction. #146
+
 ## 5.1.0
 
 - Fixes extension activation issue